+++ src/main/java/com/search/cat_picture/service/CatPictureService.java
@@ -1,38 +1,16 @@
 package com.search.cat_picture.service;
 
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import com.search.cat_picture.dto.CatPictureResponse;
 import com.search.cat_picture.dto.CatPictureResponses;
-import com.search.cat_picture.global.openFeign.theCat.TheCatClient;
 
-import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+public interface CatPictureService {
 
-@Service
-@Slf4j
-@AllArgsConstructor
-@Transactional(readOnly = true)
-public class CatPictureService {
+	 void getCatPictureBy(String id);
 
-	private final TheCatClient theCatClient;
+	 CatPictureResponses getRandomCatPictures(int limit);
 
-	public void getCatPictureBy(String id) {
-		var res = theCatClient.findCatPictureById(id);
-		log.info(res.toString());
-	}
+	 CatPictureResponses getCatPicturesByBreedName(String breedName);
 
-	public CatPictureResponses getRandomCatPictures(int limit) {
-		theCatClient.findRandomPictures(limit);
-		return null;
-	}
+	 CatPictureResponse getCatPictureById(String imageId);
 
-	public CatPictureResponses getCatPicturesByBreedName(String breedName) {
-		return null;
-	}
-
-	public CatPictureResponse getCatPictureById(String imageId) {
-		return null;
-	}
 }
